feat: add collections module for shared list,map functionalities
